pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dadaptbr"
version = "0.1.0"
description = "DADA-PT-BR"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.11.0",
    "datasets>=4.1.1",
    "matplotlib>=3.5.0",
    "numpy>=1.21.0",
    "ollama>=0.6.0",
    "psutil>=7.1.0",
    "python-dotenv>=1.1.1",
    "seaborn>=0.11.0",
    "tqdm>=4.64.0",
    "unbabel-comet>=2.2.7",
]

[project.scripts]
dada = "dadaptbr.main:main"

[tool.setuptools.packages.find]
where = ["src"]

# Black configuration
[tool.black]
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Ruff configuration
[tool.ruff]
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-first-party = ["config"]

# Flake8 configuration
[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "E501", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
    "*.egg",
]

# Pylint configuration
[tool.pylint.messages_control]
max-line-length = 88
